# docs/swagger.yaml
basePath: /
definitions:
  event.AlertThresholds:
    properties:
      critical:
        type: integer
      red:
        type: integer
      yellow:
        type: integer
    type: object
  event.ColorConfig:
    properties:
      primary:
        type: string
      secondary:
        type: string
    type: object
  event.DataConfig:
    properties:
      name:
        type: string
      unit:
        type: string
    type: object
  event.DataPoint:
    properties:
      name:
        type: string
      value: {}
    type: object
  event.Entry:
    properties:
      data:
        items:
          $ref: '#/definitions/event.DataPoint'
        type: array
      endTime:
        type: string
      eventId:
        type: string
      note:
        type: string
      receiverId:
        type: string
      startTime:
        type: string
      type:
        type: string
      userId:
        type: string
    type: object
  event.EventConfig:
    properties:
      color:
        $ref: '#/definitions/event.ColorConfig'
      data:
        $ref: '#/definitions/event.DataConfig'
      fields:
        items:
          $ref: '#/definitions/event.FieldConfig'
        type: array
      graph:
        $ref: '#/definitions/event.GraphConfig'
      hasQuickAdd:
        type: boolean
      icon:
        type: string
      monitor:
        $ref: '#/definitions/event.MonitorConfig'
      type:
        type: string
      upcoming:
        $ref: '#/definitions/event.UpcomingConfig'
    type: object
  event.FieldConfig:
    properties:
      inputType:
        type: string
      label:
        type: string
      name:
        type: string
      placeholder:
        type: string
      required:
        type: boolean
    type: object
  event.GraphConfig:
    properties:
      title:
        type: string
      type:
        type: string
    type: object
  event.MonitorConfig:
    properties:
      alertThresholds:
        $ref: '#/definitions/event.AlertThresholds'
      showLastValue:
        type: boolean
    type: object
  event.UpcomingConfig:
    properties:
      lookAheadDays:
        type: integer
      show:
        type: boolean
    type: object
...
  handlers.ReceiverEventRequest:
    properties:
      data:
        items:
          $ref: '#/definitions/event.DataPoint'
        type: array
      endTime:
        type: string
      note:
        type: string
      receiverId:
        type: string
      startTime:
        type: string
      type:
        type: string
      userId:
        type: string
    required:
    - endTime
    - receiverId
    - startTime
    - type
    - userId
    type: object
  handlers.ReceiverEventResponse:
    properties:
      eventId:
        type: string
      receiverId:
        type: string
      status:
        type: string
    type: object
...
host: api.caretosher.com
info:
  contact: {}
  description: REST API for the CareToSher care-giving collaboration platform. All
    routes require a Cognito JWT except POST /user.
  title: CareToSher API
  version: "1.0"
paths:
  /event:
    post:
      parameters:
      - description: Event details. startTime and endTime must be RFC3339. type must
          match a valid event config.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ReceiverEventRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReceiverEventResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "403":
          description: Access denied
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a care event for a receiver
      tags:
      - events
  /event/{eventId}:
    delete:
      parameters:
      - description: 'Event ID (format: Event#<uuid>)'
        in: path
        name: eventId
        required: true
        type: string
      - description: 'Receiver ID (format: Receiver#<uuid>)'
        in: query
        name: receiverId
        required: true
        type: string
      - description: 'ID of the requesting user (format: User#<uuid>)'
        in: query
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "403":
          description: Access denied
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a care event
      tags:
      - events
  /events/{receiverId}:
    get:
      parameters:
      - description: 'Receiver ID (format: Receiver#<uuid>)'
        in: path
        name: receiverId
        required: true
        type: string
      - description: 'ID of the requesting user (format: User#<uuid>)'
        in: query
        name: userId
        required: true
        type: string
      - description: Start of time range (RFC3339). Must be provided with endTime.
        in: query
        name: startTime
        type: string
      - description: End of time range (RFC3339). Must be provided with startTime.
        in: query
        name: endTime
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/event.Entry'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "403":
          description: Access denied
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all care events for a receiver
      tags:
      - events
  /events/configs:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/event.EventConfig'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all available event type configurations
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"